third_party/WebKit/LayoutTests/http/tests/misc/script-sync-slow-scripts-onerror-expected.txt

Slow scripts that fail to load should not dispatch multiple error events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstOnErrorHandlerHasRun is false
PASS firstOnErrorHandlerHasRun is true
PASS secondOnErrorHandlerHasRun is false
PASS successfullyParsed is true

TEST COMPLETE
